% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phipmake2-package.R
\docType{package}
\name{phipmake2-package}
\alias{phipmake2}
\alias{phipmake2-package}
\title{phipmake2: fast, memory-bounded PhIP-Seq summarisation}
\description{
A rewrite of the Larman Lab phipmake post-alignment stage. See
[run_phipmake2()] for the entry point, and the package README for the
rationale behind the rewrite.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/avinashkarkada/PhIPmake2}
  \item Report bugs at \url{https://github.com/avinashkarkada/PhIPmake2/issues}
}

}
\author{
\strong{Maintainer}: Avinash Karkada \email{avikarkada@gmail.com}

Authors:
\itemize{
  \item Avinash Karkada \email{avikarkada@gmail.com}
}

}
\keyword{internal}
